# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "version": "0.1"
    },
    "paths": {
        "/api/buku": {
            "get": {
                "tags": [
                    "Buku"
                ],
                "operationId": "ad144718c94735e4b410f4451f86cbf1",
                "responses": {
                    "200": {
                        "description": "Get all data buku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "judul": {
                                                "type": "string",
                                                "example": "Belajar PHP"
                                            },
                                            "penulis": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "harga": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 99.99
                                            },
                                            "tgl_terbit": {
                                                "type": "integer",
                                                "example": 2021
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2021-08-01T00:00:00.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2021-08-01T00:00:00.000000Z"
                                            },
                                            "photo": {
                                                "type": "string",
                                                "example": "buku.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Buku"
                ],
                "operationId": "17d1429cfe1ba64127b7794520e46c85",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Belajar PHP"
                                    },
                                    "penulis": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "tgl_terbit": {
                                        "type": "integer",
                                        "example": 2021
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "buku.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store data buku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data berhasil disimpan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/gallery": {
            "get": {
                "tags": [
                    "Gallery"
                ],
                "description": "Get all posts with picture",
                "operationId": "f43620079a76c2c3b2ce527fe99c3e72",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Impedit omnis vitae"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Nemo est velit sapie"
                                            },
                                            "picture": {
                                                "type": "string",
                                                "example": "posts/FZN9JEP3UYagtjMluLK8y74bMy0jWzNVzZb5Cz3i.jpg"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2024-11-05T19:07:54.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2024-11-05T19:07:54.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Gallery"
                ],
                "description": "Create new Post",
                "operationId": "a008e5cea80c123e7654800a7d4093a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "picture": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Post created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/gallery/{id}": {
            "delete": {
                "tags": [
                    "Gallery"
                ],
                "description": "Delete Post by ID",
                "operationId": "4c2127cd3850b5284aa330b0e64a6761",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Post deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "Greet"
                ],
                "summary": "Greet a user",
                "operationId": "cc6e5e6e5d9b60e2de887c36d09ca489",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "John",
                                "Jane"
                            ]
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Buku",
            "description": "Buku"
        },
        {
            "name": "Gallery",
            "description": "Gallery"
        },
        {
            "name": "Greet",
            "description": "Greet"
        }
    ]
}